Add progress for auth-config
