+++ test/unit/auth/tenant.spec.ts
@@ -22,7 +22,7 @@ import * as chaiAsPromised from 'chai-as-promised';
 import { deepCopy } from '../../../src/utils/deep-copy';
 import {
   EmailSignInConfig, EmailSignInProviderConfig, MultiFactorAuthConfig,
-} from '../../../src/auth/auth-config';
+} from '../../../src/auth/auth-config-internal';
 import {
   Tenant, TenantOptions, TenantServerResponse,
 } from '../../../src/auth/tenant';
